+++ src/SqlArtisan/Internal/SqlPart/Keywords.cs
@@ -70,6 +70,7 @@ internal static class Keywords
     internal const string Group = "GROUP";
     internal const string GroupConcat = "GROUP_CONCAT";
     internal const string Grouping = "GROUPING";
+    internal const string GroupingId = "GROUPING_ID";
     internal const string Having = "HAVING";
     internal const string In = "IN";
     internal const string Inner = "INNER";